PHT refactor
class parsing
