+++ Sources/PureHeaderTool/PureHeaderGenerator/Private/Parser/SFileData.cpp
@@ -1,0 +1,146 @@
+
+#include "Parser/SFileData.h"
+#include <fstream>
+#include <sstream>
+#include "Parser/STypes.h"
+#include "Parser/SObject.h"
+#include "Utils/Utils.h"
+
+using namespace Parser;
+
+Parser::SFileReference::SFileReference(const std::filesystem::path& inFilePath, const std::filesystem::path& inReflectedPath)
+	: FilePath(inFilePath), ReflectedPath(inReflectedPath) {
+	std::filesystem::file_time_type file_time = std::filesystem::last_write_time(FilePath);
+	LastEdit = std::chrono::system_clock::to_time_t(
+		std::chrono::time_point_cast<std::chrono::system_clock::duration>(
+			file_time - std::filesystem::file_time_type::clock::now() + std::chrono::system_clock::now()
+			)
+	);
+}
+
+bool Parser::SFileReference::IsUpToDate() const {
+	std::ifstream FileStream(ReflectedPath);
+	char* LineData = new char[1000];
+	FileStream.getline(LineData, 1000);
+	String LineString(LineData);
+	delete LineData;
+	return LineString.IsStartingWith(FormatDate(LastEdit));
+}
+
+const String Parser::SFileReference::GetName() const {
+	return FilePath.filename().u8string().c_str();
+}
+
+String Parser::SFileReference::FormatDate(time_t inTime) {
+	std::tm* gmt = new std::tm();
+#if _WIN32
+	gmtime_s(gmt, &inTime);
+#else
+	gmtime_r(&inTime, gmt);
+#endif
+
+	std::stringstream TimeBuffer;
+	TimeBuffer << std::put_time(gmt, "%A, %d %B %Y %H:%M:%S");
+	delete gmt;
+	return String("//LAST UPDATE : ") + TimeBuffer.str().c_str();
+}
+
+Parser::SFileData::SFileData(const SFileReference& inFilePath)
+	: File(inFilePath) {
+
+	std::ifstream FileStream(File.GetFilePath());
+	Content = "";
+	std::string line;
+	while (std::getline(FileStream, line)) {
+		Content << line.c_str() << '\n';
+	}
+}
+
+void Parser::SFileData::SFileData::ParseContent() {
+	SStateMachine Status;
+
+	String Simplified = Content;
+
+
+	for (Status.CurrentPos = 0; Status.CurrentPos < Simplified.Length(); ++Status.CurrentPos) {
+		const char* currentData = &Simplified.GetData()[Status.CurrentPos];
+
+		if (Status.DoesSkipChar(currentData)) continue;
+
+		if (!Status.IsReadingStructure) {
+			if (Utils::IsStartingWith(currentData, "REFLECT(")) {
+				Status.IsReadingStructure = true;
+				Status.Content = "";
+			}
+		}
+
+		if (Status.IsReadingStructure) {
+			Status.Content << currentData[0];
+			if (currentData[0] == '{') {
+				Status.IndentationLevel++;
+			}
+			else if (currentData[0] == '}') {
+				Status.IndentationLevel--;
+				if (Status.IndentationLevel == 0) {
+					ParseStructureHeader(Status);
+					Status.IsReadingStructure = false;
+				}
+			}
+		}
+	}
+}
+
+void Parser::SFileData::ParseStructureHeader(const SStateMachine& Structure) {
+	SStateMachine Status;
+	EObjectType ObjectType = EObjectType::ObjType_None;
+	for (int i = 0; i < Structure.Content.Length(); ++i) {
+		const char* CurrentData = &Structure.Content.GetData()[i];
+
+		if (CurrentData[0] == '{') break;
+
+		//Extract structure type
+		if (ObjectType == EObjectType::ObjType_None) {
+			if (Utils::IsStartingWith(CurrentData, "enum")) ObjectType = EObjectType::ObjType_Enum;
+			if (Utils::IsStartingWith(CurrentData, "class")) ObjectType = EObjectType::ObjType_Struct;
+			if (Utils::IsStartingWith(CurrentData, "struct"))  ObjectType = EObjectType::ObjType_Class;
+			continue;
+		}
+
+		//Then extract name
+		if (CurrentData[0] == ' ' || CurrentData[0] == '\t' || CurrentData[0] == '\n' || CurrentData[0] == '}') {
+			Status.PauseCapture = true;
+		}
+		else if (Status.PauseCapture) {
+			if (Utils::IsStartingWith(CurrentData, "final")) {
+				i += String("final").Length() - 1;
+				continue;
+			}
+			Status.Content = "";
+			Status.PauseCapture = false;
+		}
+		if (!Status.PauseCapture) Status.Content << CurrentData[0];
+	}
+
+	SObject* NewObject = nullptr;
+	switch (ObjectType)
+	{
+	case Parser::EObjectType::ObjType_Enum:
+		NewObject = new SEnum(Status.Content, ObjectType);
+		break;
+	case Parser::EObjectType::ObjType_Struct:
+		NewObject = new SStruct(Status.Content, ObjectType);
+		break;
+	case Parser::EObjectType::ObjType_Class:
+		NewObject = new SStruct(Status.Content, ObjectType);
+		break;
+	}
+
+	if (!NewObject) return;
+
+	NewObject->ParseContent(Structure.Content);
+	Objects.push_back(NewObject);
+}
+
+Parser::SFileData::~SFileData() {
+	for (auto& Object : Objects) delete Object;
+}
