+++ Sources/PureHeaderTool/PureHeaderGenerator/Public/Parser/STypes.h
@@ -1,0 +1,22 @@
+#pragma once
+#include "Types/String.h"
+
+namespace Parser {
+	enum class EObjectType {
+		ObjType_None = 0,
+		ObjType_Enum = 1,
+		ObjType_Struct = 2,
+		ObjType_Class = 3
+	};
+
+	struct SPropertyData {
+		String PropertyType;
+		String PropertyName;
+	};
+
+	struct SFunctionData {
+		String ReturnType;
+		String FunctionName;
+		std::vector<SPropertyData> Parameters;
+	};
+}
