+++ Sources/PureHeaderTool/PureHeaderGenerator/Public/Utils/Utils.h
@@ -6,4 +6,8 @@ namespace Utils {
 
 	void Log(const String& message);
 	void Ensure(bool condition, const String& message);
+
+	bool IsStartingWith(const char* Data, const char* Start);
+
+	bool IsVoidChar(const char Char);
 }
